Add new methods
Specifically add a method to make the current buffer a scratch buffer.
Also added a method that allows the registration of a vim event (using
autocmd) to a python function. Defined two example events as an example
of how this will work - so far this is just an example.

Also added the outline of a method to register a completion method. This
method will eventually take care of setting up the completion and
returning the correct parameters